Add files via upload
removes rain and fog.
to install this. just replace the current 1 u have in server/db. however make a back up in case of any errors.
i have not had enough time to fully test this i did it at 4am. got sick of dying to people i couldnt see and or hear.